tools.py
import PyPDF2
import docx
import requests
from bs4 import BeautifulSoup

def extract_text_from_pdf(file_path):
    with open(file_path, 'rb') as file:
        reader = PyPDF2.PdfFileReader(file)
        text = ''
        for page_num in range(reader.numPages):
            text += reader.getPage(page_num).extract_text()
        return text

def extract_text_from_docx(file_path):
    doc = docx.Document(file_path)
    return '\n'.join([para.text for para in doc.paragraphs])

def fetch_job_description(job_link):
    response = requests.get(job_link)
    soup = BeautifulSoup(response.content, 'html.parser')
    # Implement logic for extracting job description from the HTML content
    job_description = soup.get_text()  # Placeholder: Adjust to extract specific job info
    return job_description

def research(query):
    url = "https://google.serper.dev/search"
    payload = {"q": query}
    headers = {
        'X-API-KEY': os.getenv('SERPER_API_KEY'),
        'Content-Type': 'application/json'
    }
    response = requests.post(url, headers=headers, json=payload)
    return response.json()

def write_content(research_material, topic):
    # Implement writing function
    pass
