requirements.txt
openai
autogen
python-docx
PyPDF2
requests
beautifulsoup4
dotenv